# Conflicts:
#	feature/notification/src/main/java/com/teamwable/notification/action/NotificationActionFragment.kt

+++ core/network/src/main/java/com/teamwable/network/dto/response/notification/ResponseNotificationsDto.kt
@@ -26,7 +26,7 @@ data class ResponseNotificationsDto(
     @SerialName("isDeleted")
     val isDeleted: Boolean = false,
     @SerialName("notificationId")
-    val notificationId: Int = -1,
+    val notificationId: Long = -1,
     @SerialName("triggerMemberId")
     val triggerMemberId: Int = -1,
 )
